fix: move some chat fn out of agents code
