+++ packages/server/src/api/controllers/ai/index.ts
@@ -5,3 +5,4 @@ export * from "./js"
 export * from "./tables"
 export * from "./configs"
 export * from "./chatApps"
+export * from "./chatConversations"
